Get expiration on traces "working"
Extremely unclean code, but this gets my pcap values expiring. TBD if
it's fast enough to loop through all of the messages each time Expire is
called, but if that's ok, this isn't too bad. And it uses NATS to its
fullest, since it may also be nice to have a separate feed for this.